Add modeled exception handling to XmlErrorParser, remove RestXmlErrorParser
