+++ src/Api/ErrorParser/XmlErrorParser.php
@@ -4,6 +4,7 @@ namespace Aws\Api\ErrorParser;
 use Aws\Api\Parser\PayloadParserTrait;
 use Aws\Api\Parser\XmlParser;
 use Aws\Api\Service;
+use Aws\Api\StructureShape;
 use Aws\CommandInterface;
 use Psr\Http\Message\ResponseInterface;
 
@@ -13,6 +14,8 @@ use Psr\Http\Message\ResponseInterface;
 class XmlErrorParser extends AbstractErrorParser
 {
     use PayloadParserTrait;
+
+    protected $parser;
 
     public function __construct(Service $api = null, XmlParser $parser = null)
     {
@@ -39,6 +42,44 @@ class XmlErrorParser extends AbstractErrorParser
             $this->parseBody($this->parseXml($body, $response), $data);
         } else {
             $this->parseHeaders($response, $data);
+        }
+
+        if (!empty($command) && !empty($this->api)) {
+
+            // If modeled error code is indicated, check for known error shape
+            if (!empty($data['code'])) {
+
+                $errors = $this->api->getOperation($command->getName())->getErrors();
+                foreach ($errors as $key => $error) {
+
+                    // If error code matches a known error shape, populate the body
+                    if ($data['code'] == $error['name']
+                        && $error instanceof StructureShape
+                    ) {
+                        $modeledError = $error;
+                        $data['body'] = $this->extractPayload(
+                            $modeledError,
+                            $response
+                        );
+
+                        foreach ($error->getMembers() as $name => $member) {
+                            switch ($member['location']) {
+                                case 'header':
+                                    $this->extractHeader($name, $member, $response, $data['body']);
+                                    break;
+                                case 'headers':
+                                    $this->extractHeaders($name, $member, $response, $data['body']);
+                                    break;
+                                case 'statusCode':
+                                    $this->extractStatus($name, $response, $data['body']);
+                                    break;
+                            }
+                        }
+
+                        break;
+                    }
+                }
+            }
         }
 
         return $data;
@@ -93,5 +134,18 @@ class XmlErrorParser extends AbstractErrorParser
         // being able to handle it :(.
         $element->registerXPathNamespace('ns', $namespaces['']);
         return 'ns:';
+    }
+
+    protected function payload(
+        ResponseInterface $response,
+        StructureShape $member
+    ) {
+        $xmlBody = $this->parseXml($response->getBody(), $response);
+        $prefix = $this->registerNamespacePrefix($xmlBody);
+        $errorBody = $xmlBody->xpath("//{$prefix}Error");
+
+        if (is_array($errorBody) && !empty($errorBody[0])) {
+            return $this->parser->parse($member, $errorBody[0]);
+        }
     }
 }
