+++ tests/Api/ErrorParser/XmlErrorParserTest.php
@@ -10,62 +10,294 @@ use PHPUnit\Framework\TestCase;
  */
 class XmlErrorParserTest extends TestCase
 {
+    use ErrorParserTestServiceTrait;
+
     /**
-     * @return array
+     * @dataProvider errorResponsesProvider
+     *
+     * @param $response
+     * @param $command
+     * @param $parser
+     * @param $expected
      */
-    public function xmlDataProvider()
+    public function testParsesClientErrorResponses(
+        $response,
+        $command,
+        $parser,
+        $expected
+    ) {
+        $response = Psr7\parse_response($response);
+        $result = $parser($response, $command);
+        $this->assertArraySubset($expected, $result);
+        $this->assertInstanceOf('SimpleXMLElement', $result['parsed']);
+    }
+
+    public function errorResponsesProvider()
     {
+        $ec2Service = $this->generateTestService('ec2');
+        $ec2Client = $this->generateTestClient($ec2Service);
+        $ec2Command = $ec2Client->getCommand('TestOperation', []);
+
+        $queryService = $this->generateTestService('query');
+        $queryClient = $this->generateTestClient($queryService);
+        $queryCommand = $queryClient->getCommand('TestOperation', []);
+
+        $restXmlService = $this->generateTestService('query');
+        $restXmlClient = $this->generateTestClient($restXmlService);
+        $restXmlCommand = $restXmlClient->getCommand('TestOperation', []);
+
         return [
+            // ec2, modeled exception
             [
+                "HTTP/1.1 400 Bad Request\r\n" .
+                "TestHeader: foo-header\r\n" .
+                "x-meta-foo: foo-meta\r\n" .
+                "x-meta-bar: bar-meta\r\n" .
+                "x-amzn-requestid: xyz\r\n\r\n" .
                 '<?xml version="1.0" encoding="UTF-8"?>' . "\n" .
                 '<Response>' .
                 '  <Errors>' .
                 '    <Error>' .
-                '      <Code>Error code text</Code>' .
-                '      <Message>Error message</Message>' .
+                '      <Code>TestException</Code>' .
+                '      <Message>Error Message</Message>' .
+                '      <TestString>SomeString</TestString>' .
+                '      <TestInt>456</TestInt>' .
                 '    </Error>' .
                 '  </Errors>' .
-                '  <RequestID>xyz</RequestID>' .
-                '</Response>'
-            ],
-            [
-                '<?xml version="1.0" encoding="UTF-8"?>' . "\n" .
-                '<Error>' .
-                '  <Code>Error code text</Code>' .
-                '  <Message>Error message</Message>' .
-                '  <Resource>Foo</Resource>' .
                 '  <RequestId>xyz</RequestId>' .
-                '</Error>'
+                '</Response>',
+                $ec2Command,
+                new XmlErrorParser($ec2Service),
+                [
+                    'type' => 'client',
+                    'request_id' => 'xyz',
+                    'code' => 'TestException',
+                    'message' => 'Error Message',
+                    'body' => [
+                        'TestHeaderMember' => 'foo-header',
+                        'TestHeaders' => [
+                            'foo' => 'foo-meta',
+                            'bar' => 'bar-meta',
+                        ],
+                        'TestStatus' => 400,
+                        'TestString' => 'SomeString',
+                        'TestInt' => 456
+                    ],
+                ],
             ],
+            // ec2, no modeled shape
             [
-                '<ErrorResponse xmlns="https://sts.amazonaws.com/doc/2011-06-15/">' .
+                "HTTP/1.1 400 Bad Request\r\n" .
+                "TestHeader: foo-header\r\n" .
+                "x-meta-foo: foo-meta\r\n" .
+                "x-meta-bar: bar-meta\r\n" .
+                "x-amzn-requestid: xyz\r\n\r\n" .
+                '<?xml version="1.0" encoding="UTF-8"?>' . "\n" .
+                '<Response>' .
+                '  <Errors>' .
+                '    <Error>' .
+                '      <Code>NonExistentException</Code>' .
+                '      <Message>Error Message</Message>' .
+                '      <TestString>SomeString</TestString>' .
+                '      <TestInt>456</TestInt>' .
+                '    </Error>' .
+                '  </Errors>' .
+                '  <RequestId>xyz</RequestId>' .
+                '</Response>',
+                $ec2Command,
+                new XmlErrorParser($ec2Service),
+                [
+                    'type' => 'client',
+                    'request_id' => 'xyz',
+                    'code' => 'NonExistentException',
+                    'message' => 'Error Message',
+                ],
+            ],
+            // query, modeled exception
+            [
+                "HTTP/1.1 400 Bad Request\r\n" .
+                "TestHeader: foo-header\r\n" .
+                "x-meta-foo: foo-meta\r\n" .
+                "x-meta-bar: bar-meta\r\n" .
+                "x-amzn-requestid: xyz\r\n\r\n" .
+                '<ErrorResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/">' .
                 '  <Error>' .
-                '    <Type>Error code text</Type>' .
-                '    <Code>Error code text</Code>' .
-                '    <Message>Error message</Message>' .
+                '    <Type>ErrorType</Type>' .
+                '    <Code>TestException</Code>' .
+                '    <Message>Error Message</Message>' .
+                '    <TestString>SomeString</TestString>' .
+                '    <TestInt>456</TestInt>' .
                 '  </Error>' .
                 '  <RequestId>xyz</RequestId>' .
-                '</ErrorResponse>'
-            ]
+                '</ErrorResponse>',
+                $queryCommand,
+                new XmlErrorParser($queryService),
+                [
+                    'type' => 'client',
+                    'request_id' => 'xyz',
+                    'code' => 'TestException',
+                    'message' => 'Error Message',
+                    'body' => [
+                        'TestHeaderMember' => 'foo-header',
+                        'TestHeaders' => [
+                            'foo' => 'foo-meta',
+                            'bar' => 'bar-meta',
+                        ],
+                        'TestStatus' => 400,
+                        'TestString' => 'SomeString',
+                        'TestInt' => 456
+                    ],
+                ],
+            ],
+            // query, no modeled shape
+            [
+                "HTTP/1.1 400 Bad Request\r\n" .
+                "TestHeader: foo-header\r\n" .
+                "x-meta-foo: foo-meta\r\n" .
+                "x-meta-bar: bar-meta\r\n" .
+                "x-amzn-requestid: xyz\r\n\r\n" .
+                '<ErrorResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/">' .
+                '  <Error>' .
+                '    <Type>ErrorType</Type>' .
+                '    <Code>NonExistentException</Code>' .
+                '    <Message>Error Message</Message>' .
+                '    <TestString>SomeString</TestString>' .
+                '    <TestInt>456</TestInt>' .
+                '  </Error>' .
+                '  <RequestId>xyz</RequestId>' .
+                '</ErrorResponse>',
+                $queryCommand,
+                new XmlErrorParser($queryService),
+                [
+                    'type' => 'client',
+                    'request_id' => 'xyz',
+                    'code' => 'NonExistentException',
+                    'message' => 'Error Message',
+                ],
+            ],
+            // rest-xml, modeled exception
+            [
+                "HTTP/1.1 400 Bad Request\r\n" .
+                "TestHeader: foo-header\r\n" .
+                "x-meta-foo: foo-meta\r\n" .
+                "x-meta-bar: bar-meta\r\n" .
+                "x-amzn-requestid: xyz\r\n\r\n" .
+                '<ErrorResponse xmlns="http://cloudfront.amazonaws.com/doc/2016-09-07/">' .
+                '  <Error>' .
+                '    <Type>ErrorType</Type>' .
+                '    <Code>TestException</Code>' .
+                '    <Message>Error Message</Message>' .
+                '    <TestString>SomeString</TestString>' .
+                '    <TestInt>456</TestInt>' .
+                '  </Error>' .
+                '  <RequestId>xyz</RequestId>' .
+                '</ErrorResponse>',
+                $restXmlCommand,
+                new XmlErrorParser($restXmlService),
+                [
+                    'type' => 'client',
+                    'request_id' => 'xyz',
+                    'code' => 'TestException',
+                    'message' => 'Error Message',
+                    'body' => [
+                        'TestHeaderMember'  => 'foo-header',
+                        'TestHeaders'       => [
+                            'foo' => 'foo-meta',
+                            'bar' => 'bar-meta',
+                        ],
+                        'TestStatus' => 400,
+                        'TestString' => 'SomeString',
+                        'TestInt' => 456
+                    ],
+                ],
+            ],
+            // rest-xml, no modeled shape
+            [
+                "HTTP/1.1 400 Bad Request\r\n" .
+                "TestHeader: foo-header\r\n" .
+                "x-meta-foo: foo-meta\r\n" .
+                "x-meta-bar: bar-meta\r\n" .
+                "x-amzn-requestid: xyz\r\n\r\n" .
+                '<ErrorResponse xmlns="http://cloudfront.amazonaws.com/doc/2016-09-07/">' .
+                '  <Error>' .
+                '    <Type>ErrorType</Type>' .
+                '    <Code>NonExistentException</Code>' .
+                '    <Message>Error Message</Message>' .
+                '    <TestString>SomeString</TestString>' .
+                '    <TestInt>456</TestInt>' .
+                '  </Error>' .
+                '  <RequestId>xyz</RequestId>' .
+                '</ErrorResponse>',
+                $restXmlCommand,
+                new XmlErrorParser($restXmlService),
+                [
+                    'type' => 'client',
+                    'request_id' => 'xyz',
+                    'code' => 'NonExistentException',
+                    'message' => 'Error Message',
+                ],
+            ],
+            // legacy format, modeled exception
+            [
+                "HTTP/1.1 400 Bad Request\r\n" .
+                "TestHeader: foo-header\r\n" .
+                "x-meta-foo: foo-meta\r\n" .
+                "x-meta-bar: bar-meta\r\n" .
+                "x-amzn-requestid: xyz\r\n\r\n" .
+                '<Error>' .
+                '  <Type>ErrorType</Type>' .
+                '  <Code>TestException</Code>' .
+                '  <Message>Error Message</Message>' .
+                '  <RequestId>xyz</RequestId>' .
+                '  <Resource>Foo</Resource>' .
+                '  <TestString>SomeString</TestString>' .
+                '  <TestInt>456</TestInt>' .
+                '</Error>',
+                $restXmlCommand,
+                new XmlErrorParser($restXmlService),
+                [
+                    'type' => 'client',
+                    'request_id' => 'xyz',
+                    'code' => 'TestException',
+                    'message' => 'Error Message',
+                    'body' => [
+                        'TestHeaderMember'  => 'foo-header',
+                        'TestHeaders'       => [
+                            'foo' => 'foo-meta',
+                            'bar' => 'bar-meta',
+                        ],
+                        'TestStatus' => 400,
+                        'TestString' => 'SomeString',
+                        'TestInt' => 456
+                    ],
+                ],
+            ],
+            // legacy format, no modeled shape
+            [
+                "HTTP/1.1 400 Bad Request\r\n" .
+                "TestHeader: foo-header\r\n" .
+                "x-meta-foo: foo-meta\r\n" .
+                "x-meta-bar: bar-meta\r\n" .
+                "x-amzn-requestid: xyz\r\n\r\n" .
+                '<Error>' .
+                '  <Type>ErrorType</Type>' .
+                '  <Code>NonExistentException</Code>' .
+                '  <Message>Error Message</Message>' .
+                '  <RequestId>xyz</RequestId>' .
+                '  <Resource>Foo</Resource>' .
+                '  <TestString>SomeString</TestString>' .
+                '  <TestInt>456</TestInt>' .
+                '</Error>',
+                $restXmlCommand,
+                new XmlErrorParser($restXmlService),
+                [
+                    'type' => 'client',
+                    'request_id' => 'xyz',
+                    'code' => 'NonExistentException',
+                    'message' => 'Error Message',
+                ],
+            ],
         ];
-    }
-
-    /**
-     * @dataProvider xmlDataProvider
-     */
-    public function testParsesResponses($xml)
-    {
-        $response = Psr7\parse_response(
-            "HTTP/1.1 400 Bad Request\r\n\r\n{$xml}"
-        );
-        $parser = new XmlErrorParser();
-        $result = $parser($response);
-        $this->assertInternalType('array', $result);
-        $this->assertEquals('client', $result['type']);
-        $this->assertEquals('Error code text', $result['code']);
-        $this->assertEquals('Error message', $result['message']);
-        $this->assertEquals('xyz', $result['request_id']);
-        $this->assertInstanceOf('SimpleXMLElement', $result['parsed']);
     }
 
     public function testParsesResponsesWithNoBodyAndNoRequestId()
